refactor : 불필요한 HomeCardContainer 제거 후 Home으로 이동
